CBECC source (1079) consistent w/ CBECC-Com & -Res 2019.1.0 releases May/19 All projects and modules modified to support 64-bit executables for CBECC-Com (while -Res remains 32-bit) BEMProc updates (1079): - revise Spc daylit area error check to allow for daylit area sum to exceed fSpcFlrArea by as much as 1% (Com tic #2935) - mods to CreateIAQRptObjects() to allow for 2019 Res single family ADU reporting (Res tic #1010) - increased BEM_MAX_COMP_ID from 214 to 300 to accommodate Res CF1R XML schema objects - added BEMPX_WriteProjectFile() argument to toggle off reporting of enumeration errors (for mid-analysis Res input files) - fix bug to prevent property label 'afterchildren_' from being written to XML for array properties (Res tic #1114) - new feature to write empty/undefined properties flagged as 'primary' to XML files BEMCmpMgr updates (1079): - addition of EPlusRunMgr classes and related code to manage EnergyPlus simulations during Com analysis, including upgrades to progress reporting by simulation month - mods to facilitate export of mid-analysis Res input files w/ extension .ribd##i - allow for removal of all PVArray & Batt from Com models except when user specifies their inclusion (Com tic #2939) - enable individual Zone Clg/Htg bypass of UMLH limits (Com tic #2680) - added evaluation of 'AnalysisPostProcessing' rulelist to Com analysis for CalGreen calcs (Com tic #2925) - added WeatherFileKey override of Res (live & batch) analysis weather file (for CEC research) - removed eval of ProposedInput rules during storage of results to Res model following analysis to avoid unintended changes to model - major overhaul of CSV format, eliminating many unused columns, improving on the organization and consolidating C02-reporting (Res tic #1053) OS_Wrap updates (1079):   (mods related to upgrade of E+ from 8.5 to 9.0) - replace references to OpenStudio's Meter w/ OutputMeter - remove references (and related code) to OpenStudio's RunManager - added call to close sqlFile following results read (partially fixed analysis file clean-up) - mods to accommodate IDD mods to RunPeriod object CBECCUI updates (1079): - addition of PVArray & Battery objects to Com Mechanical object tree - inclusion of BEMGrid in Res app w/ additions to accommodate new Duct Design (DuctSeg) UI (object creation & deletion buttons, display of children of single object) - fix bug where 2016 user models imported into 2019 tool not fully switched over to 2019 (RulesetSymVal) - added WeatherFileKey & BypassMessageBoxes Res batch processing options - fix proper results display for New vs. EAA Res analysis (no EDRs in EAA)

+++ src/BEMCmpMgr/BEMCmpMgrSecure.cpp
@@ -56,7 +56,7 @@ static char THIS_FILE[] = __FILE__;
 bool CMX_SecureExecutable()
 {	char test[] = "Check String";
 	QString qsTest = test;
-	CM_CryptEncode( test, strlen( test ) );
+	CM_CryptEncode( test, (int) strlen( test ) );
 	return (qsTest.compare(test)!=0);
 }
